requirements.txt
ccxt>=4.4.0
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=16.0.0
pydantic>=2.6.0
PyYAML>=6.0.0
websockets>=12.0.0
tenacity>=8.2.0
python-dateutil>=2.9.0.post0
scipy>=1.10.0
scikit-learn>=1.3.0
gymnasium>=0.29.0
stable-baselines3>=2.3.0
chromadb>=0.5.0
faiss-cpu>=1.7.4

# Tests
pytest>=8.0.0
